Re-center octopus and positioned hat on top.
Used transform with translate to make sure the octopus was centered on the page. I did not know the size of the octopus. The translate value for transform is based off the size of the element. With the octopus centered, I then positioned the hat with position: absolute and some top and left modifications.