Created program to update a hash with user info
program will ask user to update information by predetermined keys. A method will then update the hash values to make sure they are the proper data type. The hash will be printed for the user to check to see if everything looks correct. If the data looks wrong, it will prompt the user to pick a key to update and enter a new value. The program will then print out the updated values.