Created program for release 5.5
Created program that takes a users name and coverts it to an alias. Vowels are changed to the next vowel in the alphabet and consonants are changed to the next consonant in the alphabet. The name is then reversed and the first and last names are capitalized. There is also a user interface that can accept multiple names, convert them and then store them in a hash.